Add batch file to sign main executables
All future releases of UIforETW will have the two main executables
signed. This is particularly important since these two are run as
administrator.

SHA1 signing is currently not used because it does not seem to add any
value for the operating systems (Windows 7 and above, and maybe Windows
Vista, but probably not) that UIforETW targets. If SHA1 is needed I can
add it - but I hate having to type the signing-key password twice.