Update packaging script to sign and create .cab
The package_etw.bat script now signs UIforETW.exe and UIforETW32.exe.
It then puts the .zip file in a .cab file so that the .cab file can
be signed. This adds an extra step to downloading UIforETW, but
should avoid Smart Screen problems and should allow greater trust.

Having the two binaries that require elevation now signed is obviously
a good thing.