Many updates to the GeneralTracing lab setup
This changes the pipeline of the lab tracing to make it more lab ready,
and more efficient.

A pair of python scripts are used to create a .json file that summarizes
the results.

CreateExporterConfig.py creates a .json file that can be passed to
wpaexporter so that all the .csv reports can be generated at once. This
script uses xperf -i trace.etl -a marks to more efficiently get the
location of the Starting/Stopping marks. Together this means that
wpaexporter only needs to be run once, which is good because it is quite
slow to start up.

SummarizeData.py goes through the data and creates a lab-usable .json
summary of the most interesting data points. This currently consists of
14 rows of data.

The .wpaProfile files were cleaned up, fixed, expanded, and manually
trimmed of their fat.

The start_tracing.bat and stop_tracing.bat files were changed to not
rely on leaking of environment variables - instead they both call
set_vars.bat and use setlocal to ensure that they don't leak. They also
don't default to calling EventEmitter.exe - it affects the results too
much in some scenarios. They only call it if USE_EVENT_EMITTER=1