Add a detailed flag to the lab start_tracing.bat
When doing lab tracing it is useful to have two different tracing modes.
In one mode a very lightweight trace is recorded in order to get CPU
and GPU usage, and little else. This mode has minimal effect on
performance.

The second mode is a detailed mode that records call stacks on CPU
samples and context switches. This distorts results but gives an ETL
file that can be used to investigate regressions.

If 'detailed' is passed to start_tracing.bat it starts a detailed
trace. Stopping and summarizing the trace is unchanged. The trace
files created will be larger - perhaps 5-10x larger?