Add option to summarize Chrome CPU usage
The IdentifyChromeProcesses.py script is used to list Chrome processes
by type, to aid in trace analysis. This change adds an option to
give a summary of CPU usage for each process and each type of process.
This requires running a separate wpaexporter action so it takes longer.
It also clutters up the output, so it is off by default.

This behavior can be triggered in two ways. If "Chrome developer" and
"Identify Chrome CPU" are both checked in the settings dialog then the
CPU usage report will be generated whenever a trace is recorded.

Alternately, if a trace is selected then the right-click context menu
has options, under Scripts, to generate the terse or verbose reports.

Clearly there is potential for adding additional analysis in the
future.