Adding output directory support to lab scripts
The original version of the lab ETW scripts would only run if the
batch files and scripts were all in the current directory. That's a bug.
And, there was no way of specifying where the output files should go.
This change fixes the bug and allows an optional output directory to be
specified as a parameter to the stop_tracing.bat file. The default
output directory is the current directory.