Improve per-process CPU report
When there is only one process of a particular type it doesn't make
sense to list the context switches and CPU usage both for the total and
for the (one) process. This change makes it so that the information is
omitted for the line where the PID is listed. This makes many reports
much less cluttered.