Add support for disabling background monitoring
This change adds a "Background monitoring" check box to the settings
dialog. This new check box is checked by default but can be unchecked
to disable most of the background monitoring threads, in order to
minimize how disruption of power measurements for light CPU loads, and
to allow long-term tracing without excessively draining power.