Minimize to system tray
The main window shows an icon in system tray when minimized and hides itself
in the taskbar. Click on the icon in taskbar restores the main window.

Usually UIforETW works in background and doesn't require a lot of attention
but it occupies space in taskbar. It especially noticeable when a lot of
applications are run and taskbar is configured to show labels for taskbar
buttons. If an application is minimized to system tray it occupies a very
small amount of space, and even it can be hidden at all.