Update packaging script to sign all binaries
It was pointed out in a comment on ETW central:
https://randomascii.wordpress.com/2015/09/24/etw-central/#comment-27945
that some anti-virus programs were detecting malware in the UIforETW
package. This is probably because these binaries are not signed. This
change modifies the packaging script so that all of the .dll and .exe
files which I create and ship are now signed.