Handle "Unknown" process in CPU Usage Summary
When generating a report on Chrome's CPU usage the exported data will
occasionally contain a process that has "Unknown" instead of the process
name and PID. This causes the script to fail even though it is quite
likely that the unknown process is irrelevant. This change teaches the
IdentifyChromeProcesses.py script to silently ignore this error.