Miscellaneous fixes from /analyze C++ Core Checks
After enabling the C++ Core Checks for /analyze I tried to get UIforETW
to build cleanly. This commit contains the remaining issues that
were found, directly or indirectly. The changes include:

- Identifying local variables that can be tagged as const - I approve of
this warning.
- Identifying functions that can be tagged as noexcept - meh?
- Adding override in a few cases. Good, I guess, but it's MFC functions.
- Identifying classes that need their copy constructors and assignment
operators deleted. I strongly approve of this warning.
- Identifying classes that would be clearer if their move constructors
and move assignment operators were deleted. This warning is "okay" but
it should *not* be the same warning number as the previous issue, which
is orders of magnitude more serious.
- Identifying uses of C style casts. I approve of this warning.
- Removed an unnecessary cast that was there because of WIN32 weirdness.
I'm not convinced that this is a good warning because two different
typedefs may only be the same due to coincidence, or may only be the
same in some build configurations. I predict some pain from this
warning.

After lots of changes, including disabling quite a few warnings, the
warning count is down from 407 to 23. The remaining warnings are ones
that I don't want to disable because they are useful in some cases, but
I can't resolve because of MFC, or I haven't decided what to do with
them. I hope that MFC fixes are provided to clean up or allow me to
clean up most of the remaining warnings.