Various bad fixes from /analyze C++ Core Checks
After enabling the C++ Core Checks for /analyze I tried to get UIforETW
to build cleanly. This commit contains comments describing some
inaccurate warnings (one caused by an MFC flaw) and many warnings that
I had to disable, either because the changes required were too great or
because the warnings were inaccurate in too many cases.

If anybody wants to fix up the code to allow removing some of the
warning disables then go for it, but only if it really does make the
code better.