Rearrange hidden columns in default WPA profile
The Annotation, Thread Name, and Readying Process Name columns are off
by default but can be useful. There use is made easier if they appear in
a logical location when enabled. This change moves these columns. The
change was done by hand editing the .wpaProfile file, to avoid
introducing spurious changes. So, I now know that KeyColumnCount is what
determines how many columns (visible or not) are to the left of the
vertical orange bar. Good to know!

Note that Thread Name in the CPU Usage (Precise) tables is a poor choice
of names - it should be New Thread Name so that it sorts with New Thread
Id. This also avoids ambiguity - there are three threads involved in a
context switch so "Thread Name" doesn't actually mean anything.

I'm tempted to make Thread Name on by default - it works very well when
profiling Chrome, which makes use of the new Windows 10 thread naming
APIs.

This change also fixes a typo in a pair of print statements.