Add good heap analysis profile
The regular heap analysis profile has odd defaults that I don't
understand. It auto-expands all processes and heaps. It shows the Handle
column (the heap handle) by default, and it doesn't show the all
important Type column, or the Stack column. This profile has a minimal
set of high-value columns, and it changes the graphing to just show the
allocation time, instead of showing the lifetime (which gets smeared
too much when multiple allocations are selected, especially AIFO
allocations).