Print whenever heap tracing flags change
Previously UIforETW would print a message whenever heap tracing was
enabled for a particular process name. That made me feel a bit tense
since it was not obvious if/when the heap tracing registry keys were
being cleared. This change adds a function to read the registry key
so that changes can be detected.