Add ARM64 support
Upgrade ETWProviders to build with VS 2017, disable code analysis, and
add an ARM64 configuration. Also required explicitly linking with
Advapi32.lib.

Add an ARM64 configuration to UIforETW, disable the assembly language
SpinALot function, and properly decide which ETWProviders DLL to load.

Added ETWProvidersARM64.dll and ETWProvidersARM64.lib.

Updated the packaging script to sign ETWProvidersARM64.dll.

The ARM64 code has not yet been tested beyond confirming that it builds.

The x86 version of UIforETW runs just fine on ARM64 Win32 so for now the
main benefit of this change is that it gives an ARM64 version of the
ETWProviders DLL.