Add WPT directory to the path
When Windows Performance Toolkit has been installed by UIforETW for the
first time it will not be in the path. This doesn't affect UIforETW
because it runs xperf.exe with a fully specified path. However this does
affect some of the Python scripts which UIforETW runs. Adding the WPT10
directory to the path avoids this first-run problem.

See issue #106 for details.