Update StripChromeSymbols.py to new symcache format
The undocumented symcache directory structure has changed again, for the
better. The new layout is basically symbol-server compatible, which
avoids having hundreds of thousands of files in one directory. Ideally
this script would handle the old and new formats but this script is used
so rarely that I'm not sure it is worth it.

This also updates the script to handle slightly different output for the
xperf action, and adds a couple of new PDB names.

This is step one towards fixing issue #131.