Add support for v3.1.0 symcache format
ETWPackSymbols.py already understood two symcache formats, so now it
understands three. It can find the new symcache files, and preserve the
directory hierarchy when copying to the destination.

The filtering of what files are "interesting" is less reliable than
ever, so adjust that line as needed.

This finishes resolving issue #131.