Add C# version of IdentifyChromeProcesses.py
Microsoft has released a trace processing package that can be used to
analyze ETW traces using .Net. This C# program uses that to implement
IdentifyChromeProcesses in C#. In some ways this cleans up the code
and the behavior (some inconsistencies and bugs in the python script
were already made) and in some ways it is messier because more of the
work is done by the script instead of by xperf actions.

Using C# and trace processing package avoids a dependency on xperf
and wpaexporter. It also avoids uncertainty about what filename
wpaexporter will use for its results - this has changed once in the
past. It also means that traces with missing events can be fully
processes - wpaexporter can't do that.

On the downside, it means that processing some traces causes
out-of-memory failures, and there are a few inconsistencies in odd
cases.