Revert "Switch from .Net Framework 4.7.2 to 4.6.1"
This reverts commit 3d7f0c209dfdd761083429074360c3857e069fd9.

The root cause was *not* that 4.7.2 was not installed on my other
machine, and in fact changing to 4.6.1 just made things worse. The real
problem was that the Visual Studio .Net components were not installed,
but the messages sure didn't suggest that.

https://twitter.com/BruceDawson0xB/status/1151005059343216640

This change has to be reverted because after installing the .Net
components this project will *only* load if I target 4.7.2. Sigh...