Uncheck "Prefer 32-bit" to avoid OOM failures
C# programs can run as 32-bit or 64-bit with the default being Any CPU.
One could be forgiven for assuming that that would be a 64-bit process
on a 64-bit OS. However that is, by default, not true.

One could also be forgiven for thinking that if one selected the x64
platform from the toolbar that one would then get a 64-bit process.
However that assumption will only lead to tears.

In both cases you get a 32-bit process by default, for two reasons.
First is that the Any CPU project settings have "Prefer 32-bit" checked
by default. Second is that the x64 solution platform maps to the Any CPU
project setting.

Trace processing can use a lot of memory. That is okay. Now it won't OOM
fail, as long as you have enough RAM or swap.