Add a batch file to record ETW heap snapshots
ETW heap snapshots are a pretty great way of doing long-term heap
tracing. They are a little intimidating (like all things ETW) and a
little lightly documented (like all things ETW) and inconvient to use.
This batch file makes it to record a heap snapshot trace and it gives it
a sensible date/time based name and slots it in with other UIforETW
traces.

As a bonus it compresses the trace - I'm not sure why this isn't done
by default.