Make C# and Python process identification virtually identical
This change resolves a few differences in the output of the C# program
that identifies Chrome processes and the original Python version. The
two tools now give identical results (except for the new lost-events
warning) on 481 out of 485 traces tested, and the four discrepancies are
on traces with weird process-parentage data that doesn't really make
sense anyway.

The results in -c mode (showing CPU usage) remain because trace
processing reports on CPU usage differently, but those differences are
known and expected and not generally significant.