Simplify heap snapshot control
Previously one had to type any key to record a trace, and then y/n to
decide whether to exit. Now you type enter each time you want to record
a trace, and 'q' and enter when you are done. This is half as many
keystrokes for each trace that is recorded, and it is much simpler, and
less error prone.