Adding a heap snapshot profile
Heap snapshot analysis needs a completely separate view, and the default
view has multiple problems, so this change adds a reasonable heap
snapshot view. The Annotation column is not initially needed but is
harmless, and handy if you add annotations. The Stack Ref# column can be
a handy way to group by unique stacks. The Stack column is vital - I
don't know why it is off by default. The Size and Count columns are
crucial.