Use -capturestate to reliably record window focus
xperf -capturestate is needed by some providers. Sometimes it is better
to do it at the start of the trace, sometimes it is better at the end,
but always it is undocumented. For the Microsoft-Windows-Win32k provider
when tracing to a file it is just important to do it, in order to get
focus data in those traces where focus doesn't change.