Update etwheapsnapshot.bat - trace compression
This change adds trace compression to ETW heap snapshot recording. I'm
not sure how that got missed when it was originally written. On the 34
heap snapshot traces that I have lying around this saves 4.6 GB (!!!)
which seems worth doing.

This also modifies the script so that it will handle it more gracefully
if xperf.exe is not available. I don't know for sure if the resulting
(unmerged) trace will work correctly but it's worth a try.