Add script to summarize CPU usage by command line
This adds a .wpaProfile file and python script to scan a trace and
summarize CPU usage by command line. The profile contains tables for
mapping from process to command line and process to CPU usage. The
script assumes that wpaexporter has been run and it processes the
resultant .csv files. The result is an output of CPU time and
command line pairs, sorted by CPU time consumed.