Add script for enumerating Randomascii presets
The startup10.wpaProfile file which ships with UIforETW sets up a
default view which I think is a better introduction to ETW profiling.
It needs updating occasionally, and sometimes some of the presets can
get lost. This script prints all of the Randomascii presets that I have
added so that I can more easily tell if any have been removed.