Update startup profile
This updates the startup profile to include the thread name column by
default, and to include the Ready Thread In DPC column. The Ready Thread
In DPC column indicates when it was a DPC, not the readying thread,
which actually did the readying. This usually happens for timer expiry.
When this column is true the Ready Thread Process/Thread is meaningless
and in fact actively misleading. I think having this column on by
default will be helpful.

The default CPU Usage (Precise) view was also changed to use stacked
lines instead of lines, so that CPU usage from various processes is
summed up instead of being overlaid.