Change version-checking code
UIforETW checks for updated versions but previously it was doing this by
scanning the UIforETW releases page looking for a version number. It
turns that scanning complex HTML that you don't control is not very
robust. So I have changed to look at a copy of Version.h that is
committed to the tree and can be updated after a new version is shipped.

Using a copy gives control over when the new release should be
"announced."

Previous versions (1.51 or before) of UIforETW won't ever notice that
they are out of date but there is nothing that can be done about that.