Updates to IdentifyChromeProcesses.py
This change gets IdentifyChromeProcesses.py to print a message when a
trace has lost events or a time inversion. It doesn't distinguish them
but I only ever see lost events so it's not worth it to me to run the
extra tests.

This change also adds -tle and -tti to the wpaexporter call. These
options are undocumented for wpaexporter but they do work for ignoring
lost events and time inversions.