Improve trace_timer_intervals
I've been using trace_timer_intervals.bat (and its associated Python
script which does all of the heavy lifting) to understand Chrome's
performance when changing the system timer interval. See
crbug.com/927165. When doing this it is convenient to filter the
results of who is changing the timer interrupt interval to just
Chrome, so I added a -f option for that, and switched to using
argparse.

At the same time I added detection of non-existent files, because
otherwise the failure modes are quite confusing.

And I changed the output to (mostly) summarize everything on a
single line per process. This is usually more readable.