Add --tabbed option to IdentifyChromeProcesses.py
When writing reports on Chrome's CPU usage it is can be helpful to paste
the results into a spreadsheet for easier presentation or additional
analysis. This change adds a --tabbed (-t) option to print tab-separated
data.

This change also shortens the names used for the utility processes
(sub_types_by_pid) and only summarizes processes by type if there is
more than one process of that type.