Cleaning up tabbed output and finding FrameServer
This change cleans up the tabbed output (-t) from the
IdentifyChromeProcesses.py script so that it can be pasted directly into
Google sheets and not require significant cleanup. It ensures that each
line stands on its own (by printing the process type as one column) and
avoids printing many unneeded lines.

This change also adds detection of the svchost.exe FrameServer process,
since this is another OS process that can be triggered (during video
conferencing) by Chrome, so it's CPU usage can be important. This
required restructing the extra-processes code to make it cleaner.

This change also removes the explicit printing of '\r' characters. This
was added because UIforETW's ChildProcess.cpp didn't do this translation
and the Windows edit control often requires \r\n line endings. Adding
these characters in the script was always wrong because they are not
needed when running from the command prompt (I guess cmd does the
translation?).

I also replaced most " characters with ', for consistency with Google's
Python coding standards, and cleaned up a few other things.

This script now handles the case where there is no CPU Usage (Precise)
data and just prints an error message and exits.

I also added a simple test batch file that runs this script on all of my
traces so I can monitor for unhandled cases.

The C# version of this code has not been updated.