Mention DISK_IO omission when circular buffer tracing
On some machines DISK_IO tracing leads to massive event loss when doing
circular buffer tracing. This is presumably a Windows bug. UIforETW has
worked around this for a while by disabling DISK_IO when doing circular
buffer tracing, but this can cause confusion, so this adds an
informative message. This resolves issue #145.