Clarify confusing GUID-length message
The GUID parameter to RetrieveSymbols must be exactly 32 characters in
length but when it isn't the tool wouldn't tell you what the length was.
It would also print some confusing information about trimmed GUIDs
which was not always relevant.

Now it prints the size and only prints the trimming information when it
is relevant (and more clearly).