Used a GUID-named mutex to find whether UIforETW is running
UIforETW used to look for a window with the title UI for ETW to see if
it was already running, but this can accidentally find the wrong window.
Using a named mutex is more robust - safe against anything but a
malicious local attacker.

With this change UIforETW will reliably launch/not-launch based on
whether it is already running. In not-launch case it may not reliably
activate the correct window but that is a low-severity edge case.

This fixes issue #147