Find wpa.exe in the path to support store versions
There are now previous versions of WPA in the Microsoft store and these
versions are newer than the versions shipped through the SDK. This
change teaches UIforETW how to search the path to find wpa.exe. This
lets developers configure their system path so that the desired version
of wpa.exe comes earliest in the path.

This should be sufficient to finally resolve issue #118.