Fix echo command in handle-tracking batch file
The echo command in the handle-tracking batch file told readers to look
for a "Heap Snapshot" graph, which was wrong. Well, it tried to tell
readers to look there but it accidentally directed the echo command to
a newly created file called "Heap". Oops. Both of these issues are fixed
now.