Update to Windows 11 WPT and drop Windows 7
This changes UIforETW to prefer and try to install the latest version of
Window Performance Toolkit (WPT) from the Windows 11 SDK. This is
currently the same version that you can install from the Microsoft Store
as Windows Performance Analyzer (Preview).

The name of the installer changed (and it got much bigger!) so the
packaging script and startup code both changed. In addition it seemed
like time to drop Windows 7 support so the packaging script and startup
code no longer have support for Windows 7. This reduces the size of the
ETW package.