Add --collapsethreads flame-graph option
In some cases (crbug.com/1231683) multiple processes are consuming CPU
together and having a separate flame graph for each thread is not
helpful. In this case it is helpful to be able to collapse the samples
from all threads in a process together. This change does that by
removing the thread ID from the collation key and by adding a fake call
stack entry to represent the thread ID within the flame graph.