Search more widely for xperf.exe
If WPT 22H2 is installed as a stand-alone product then it is installed
to ProgramFiles rather than ProgramFilesX86 and it cannot be located
in the SDK install folder. Therefore it is necessary to look in the
likely paths to see if xperf.exe can be found. GetInstallFolder() was
created and improved to handle this, and is called after running the
installers to refresh the install path.