Trim leading/trailing whitespace from most edit controls
When pasting a user provider from Signal a trailing \n was accidentally
pasted. This was dutifully preserved and put into the xperf command-line
which then failed. This was quite confusing because the trailing \n was
invisible, and unexpected since the edit control had Multiline set to
false.

So, this change trims leading and trailing whitespace from this edit
control and others.