Clean up input hook code
This cleans up the input hooking code with better comments, use of
nullptr instead of 0, and more accurate error handling.

The "more accurate error handling" is that previously if one hook was
registered and the other one was not then the message pump would not
be run, which could cause issues.