Use appropriate profile in etw_processes_longterm.bat
etw_processes_longterm.bat would print a message when it was done
suggesting what .wpaprofile to use, but this message was inappropriate
because it was targeted at a trace that recorded CPU usage, which this
batch file does not. This change updates that message to point to a
newly create profile.