Version 160 (#172)
* Fix Python3 compatibility bug in IdentifyChromeProcesses.py

IdentifyChromeProcesses.py was using pathByBrowserPid.has_key(parentPid)
and has_key doesn't exist in the dict type in Python 3. I'm not sure how
this slipped by but apparently UIforETW on my system now launches this
script with Python 3, which is good.

I tested the script with Python 2 and 3 and it now works with both
variants, although I'm not sure supporting Python 2 is particularly
important these days.

* Update to version 1.60

This change has no effect until new UIforETW binaries ship that
reference the new randomascii/UIforETW repo. Updating the two copies of
the version string so that I can create a new release and upload it.